#17 fix: moved API request to libs
