+++ app/api/categories/[slug]/route.ts
@@ -1,43 +1,12 @@
-import { client } from "@/src/lib/client";
-import { gql } from "graphql-tag";
 import { NextRequest, NextResponse } from "next/server";
 import { GetCategoriesProps, Job } from "../../types";
+import { getCategory } from "@/src/lib/categories";
 
 export async function GET(
   _request: NextRequest,
   { params }: GetCategoriesProps
 ): Promise<NextResponse> {
-  const job: Job[] = await client
-    .query({
-      query: gql`
-        query GetJobsByCategory($slug: String!) {
-          jobCollection(where: { category: { slug: $slug } }) {
-            items {
-              slug
-              title
-              cover {
-                url
-                title
-              }
-              sys {
-                id
-              }
-              categoryCollection {
-                items {
-                  slug
-                }
-              }
-            }
-          }
-        }
-      `,
-      variables: {
-        slug: params.slug,
-      },
-    })
-    .then((response) => {
-      return response.data.jobCollection.items;
-    });
+  const job: Job[] = await getCategory(params.slug);
 
   if (!job.length) {
     return NextResponse.json(
