+++ app/page.tsx
@@ -1,6 +1,8 @@
 import { Metadata } from "next";
 import { Header } from "@/src/features/Header";
 import { JobsList } from "@/src/features/JobsList";
+import { getJobs } from "@/src/lib/jobs";
+import { Job } from "@/app/api/types";
 import { data as HeaderData } from "@/src/features/Header/Header.mocks"; // TODO: replace this with actual data
 
 export const metadata: Metadata = {
@@ -9,14 +11,10 @@ export const metadata: Metadata = {
 };
 
 export default async function Home() {
-  const { homeLogo, mobileLogo } = HeaderData;// TODO: replace this with actual data
+  const { homeLogo, mobileLogo } = HeaderData; // TODO: replace this with actual data
 
-  const getJobsList = async () => {
-    const response = await fetch(process.env.URL + `/api/jobs`);
-    const { data } = await response.json();
-    return data;
-  };
-  const jobsList = await getJobsList();
+  const jobsList: Job[] = await getJobs();
+
   return (
     <>
       <Header homeLogo={homeLogo} mobileLogo={mobileLogo} />
